restructured and added deserialization.
