--- src/main/code/de/tuchemnitz/se/exercise/core/configmanager/ConfigFileEncoder.kt
@@ -1,43 +1,0 @@
-package de.tuchemnitz.se.exercise.core.configmanager
-
-import com.mongodb.client.MongoDatabase
-import de.tuchemnitz.se.exercise.persist.AbstractCollection
-import de.tuchemnitz.se.exercise.persist.configs.EyeTrackingConfig
-import de.tuchemnitz.se.exercise.persist.configs.IConfig
-import kotlinx.serialization.json.Json
-import org.bson.BsonDocument
-import org.litote.kmongo.descending
-
-class ConfigFileEncoder(private val configCollections: ConfigManager.ConfigCollections) {
-
-    fun configFile(): String {
-        val tools = assembleAllConfigurations()
-        return Json { prettyPrint = true }.encodeToString(
-            ConfigFile.serializer(),
-            ConfigFile(
-                general = ConfigManager.generalSettings,
-                bubbleViewConfig = tools.bubbleViewConfig,
-                zoomMapsConfig = tools.zoomMapsConfig,
-                codeChartsConfig = tools.codeChartsConfig,
-                eyeTrackingConfig = tools.eyeTrackingConfig,
-                dataClientConfig = ConfigManager.dataClientConfig,
-                database = ConfigManager.databaseConfig
-            )
-        )
-    }
-
-    fun assembleAllConfigurations(): ToolConfigs {
-        return ToolConfigs(
-            codeChartsConfig = configCollections.codeChartsConfigCollection.findMostRecent(),
-            zoomMapsConfig = configCollections.zoomMapsConfigCollection.findMostRecent(),
-            // TODO
-            eyeTrackingConfig = EyeTrackingConfig(dummyVal = ""),
-            // TODO
-            bubbleViewConfig = BubbleViewConfig(filter = setOf(0F))
-        )
-    }
-
-    private fun <T : IConfig> AbstractCollection<T>.findMostRecent(): T? =
-        find(BsonDocument()).sort(descending(IConfig::savedAt))
-            .firstOrNull()
-}
