+++ src/main/code/de/tuchemnitz/se/exercise/core/configmanager/ConfigManager.kt
@@ -1,6 +1,7 @@
 package de.tuchemnitz.se.exercise.core.configmanager
 
 import com.mongodb.client.MongoDatabase
+import de.tuchemnitz.se.exercise.persist.AbstractCollection
 import de.tuchemnitz.se.exercise.persist.configs.CodeChartsConfig
 import de.tuchemnitz.se.exercise.persist.configs.EyeTrackingConfig
 import de.tuchemnitz.se.exercise.persist.configs.IConfig
@@ -9,13 +10,17 @@ import de.tuchemnitz.se.exercise.persist.configs.collections.CodeChartsConfigCol
 import de.tuchemnitz.se.exercise.persist.configs.collections.EyeTrackingConfigCollection
 import de.tuchemnitz.se.exercise.persist.configs.collections.ZoomMapsConfigCollection
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import org.bson.BsonDocument
 import org.bson.conversions.Bson
+import org.litote.kmongo.descending
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.IOException
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 import java.nio.file.Path
+import java.nio.file.Paths
 
 class ConfigManager(var configFilePath: String = "", database: MongoDatabase) {
 
@@ -30,8 +35,6 @@ class ConfigManager(var configFilePath: String = "", database: MongoDatabase) {
         ZoomMapsConfigCollection(database),
         EyeTrackingConfigCollection(database)
     )
-
-    val configFileEncoder = ConfigFileEncoder(configCollections)
 
     companion object {
         val logger: Logger = LoggerFactory.getLogger(ConfigManager::class.java)
@@ -53,7 +56,7 @@ class ConfigManager(var configFilePath: String = "", database: MongoDatabase) {
         return true
     }
 
-    fun writeFile(path: Path, content: String = configFileEncoder.configFile()) {
+    fun writeFile(path: Path, content: String = "") {
         try {
             Files.writeString(path, content, StandardCharsets.UTF_8)
         } catch (e: IOException) {
@@ -87,4 +90,37 @@ class ConfigManager(var configFilePath: String = "", database: MongoDatabase) {
         generalSettings.selectionMenuEnabled = selectionMenuEnabled
         generalSettings.configPath = configPath.toString()
     }
+
+    fun configFile(): String {
+        val tools = assembleAllConfigurations()
+        return Json { prettyPrint = true }.encodeToString(
+            ConfigFile.serializer(),
+            ConfigFile(
+                general = ConfigManager.generalSettings,
+                bubbleViewConfig = tools.bubbleViewConfig,
+                zoomMapsConfig = tools.zoomMapsConfig,
+                codeChartsConfig = tools.codeChartsConfig,
+                eyeTrackingConfig = tools.eyeTrackingConfig,
+                dataClientConfig = dataClientConfig,
+                database = databaseConfig
+            )
+        )
+    }
+
+    fun assembleAllConfigurations(): ToolConfigs {
+        return ToolConfigs(
+            codeChartsConfig = configCollections.codeChartsConfigCollection.findMostRecent(),
+            zoomMapsConfig = configCollections.zoomMapsConfigCollection.findMostRecent(),
+            // TODO
+            eyeTrackingConfig = EyeTrackingConfig(dummyVal = ""),
+            // TODO
+            bubbleViewConfig = BubbleViewConfig(filter = setOf(0F))
+        )
+    }
+
+    private fun <T : IConfig> AbstractCollection<T>.findMostRecent(): T? =
+        find(BsonDocument()).sort(descending(IConfig::savedAt))
+            .firstOrNull()
+
+    val decodedConfig = readFile(Paths.get(configFilePath))?.let { Json.decodeFromString(ConfigFile.serializer(), it) }
 }
